requirements.txt
fastapi
uvicorn[standard]
torch
Pillow
python-multipart
